disabled edgeware
we support edgeware properly in v3, but in the version of the wallet
which uses this branch the balance is decoded incorrectly.
this leads to an inflated portfolio value which is disconcerting
for our users.

in theory, by removing the chain entirely, the wallet should delete
the cached balance for this chain.